documentation

+++ src/main.cpp
@@ -6,11 +6,13 @@
 
 /**
  * Program: Rocket Payload
- * Version: 0.1
- * Date: 7 March 2019
+ * Author: Jordan Martin
+ * Version: 0.2
+ * Date Created: 5 March 2019
  * Description: Tester for rocket payloads. Also program to develop payload code.
- * Last edited by: Jordan Martin
- * Reason edited: Fixed radio comms and added more code documentation
+ * Last Edited by: Jordan Martin
+ * Last Edited: 9 March 2019
+ * Reason edited: Swiched to custom gps handler
  */
 // SoftwareSerial serial_connection(3, 2); //RX=pin 11, TX=pin 10 (for arduino... connect TX to RX and vice-versa)
 // TinyGPSPlus gps;//This is the GPS object that will pretty much do all the grunt work with the NMEA data
@@ -36,7 +38,7 @@ void loop()
   float x = 12.2, y = 17.8;
   String blah = String(x, 3) + ", " + String(y, 3);
   char toWrite[blah.length()];
-  for(int i = 0; i < blah.length(); i++) {
+  for(int i = 0; i < (int)blah.length(); i++) {
     toWrite[i] = blah.charAt(i);
   }
   radio.write(&toWrite, sizeof(toWrite));
@@ -64,7 +66,7 @@ void loop()
     gps.getGPSData(&gpsDat);
     String temp = String(gpsDat.latitude, 6) + ", " + String(gpsDat.longitude, 6); // compile into string to be used later
     char location[temp.length()]; // create location character array
-    for(int i = 0; i < temp.length(); i++) { // go through each piece of the string data
+    for(int i = 0; i < (int)temp.length(); i++) { // go through each piece of the string data
       location[i] = temp.charAt(i); // add current piece of string to char array
     }
     radio.write(&location, sizeof(location)); // send location data to receiver through the radio
